+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Note that since we don't clearly distinguish between a public and private interf
 ## [Unreleased]
 
 - Fix ugly camera clipping in snapshot transitions
+- Add viewport button to toggle illumination mode
 
 ## [v5.3.0] - 2025-11-5
 - Update loading message in MVS Stories Viewer